+++ apps/docs/theme/shadows.ts
@@ -1,0 +1,5 @@
+const shadows = {
+  outline: "0 0 0 3px #F4EBFF",
+};
+
+export default shadows;