+++ apps/starter/modules/common/components/index.ts
@@ -1,4 +1,6 @@
+export { default as ConnectWalletModal } from "./ConnectWalletModal";
 export { default as Layout } from "./Layout";
-export { default as Navbar } from "./Navbar";
+export * from "./MinimalLineChart";
+export * from "./Navbar";
+export * from "./Stat";
 export { default as TerraWallet } from "./TerraWallet";
-export { default as WalletModal } from "./WalletModal";