+++ apps/starter/pages/_app.tsx
@@ -6,9 +6,14 @@ import {
   WalletProvider,
 } from "@terra-money/wallet-provider";
 import App, { AppProps, AppContext } from "next/app";
+import { TerraProvider } from "@wizard-ui/terra";
 import { ChakraProvider, CSSReset } from "@chakra-ui/react";
+import { QueryClient, QueryClientProvider } from "react-query";
 
 import { Layout } from "modules/common";
+import theme from "../theme";
+
+const queryClient = new QueryClient();
 
 const MyApp = ({
   Component,
@@ -18,12 +23,16 @@ const MyApp = ({
 }: AppProps & WalletControllerChainOptions) => {
   const main = (
     <>
-      <ChakraProvider>
-        <CSSReset />
-        <Layout>
-          <Component {...pageProps} />
-        </Layout>
-      </ChakraProvider>
+      <QueryClientProvider client={queryClient}>
+        <ChakraProvider theme={theme}>
+          <TerraProvider>
+            <CSSReset />
+            <Layout>
+              <Component {...pageProps} />
+            </Layout>
+          </TerraProvider>
+        </ChakraProvider>
+      </QueryClientProvider>
     </>
   );