+++ packages/terra/src/index.tsx
@@ -1,5 +1,0 @@
-export * from "./message";
-export * from "./number";
-export * from "./types";
-export * from "./hooks";
-export * from "./helpers";